requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-ollama
python-dotenv